[Module] Add afterModuleBuilt hook
Add the `afterModuleBuilt` hook to `RawModule`, similar to the existing
`atModuleBodyEnd`. This hook is called after a module has been fully
constructed and its resulting component definition has been added to the
builder. Thus code inside the body of `afterModuleBuilt` can get the
definition of the surrounding module and instantiate it, which is
interesting for generating additional collateral like unit tests
alongside a module.